# api_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Clinic Management API",
    "version": "1.0.0"
  },
  "paths": {
    "/patients": {
      "post": {
        "summary": "Create a new patient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "date_of_birth": { "type": "string" },
                  "phone": { "type": "string" },
                  "address": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["name", "date_of_birth", "phone", "address", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" },
                    "access_token": { "type": "string" },
                    "token_type": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all patients (staff only)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of patients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Patient"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}": {
      "get": {
        "summary": "Get patient by ID (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Patient" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update patient (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Patient" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Patient updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Patient" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete patient (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Patient deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/doctors": {
      "post": {
        "summary": "Create a new doctor (staff only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Doctor" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Doctor created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all doctors",
        "responses": {
          "200": {
            "description": "List of doctors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Doctor" }
                }
              }
            }
          }
        }
      }
    },
    "/doctors/{id}": {
      "get": {
        "summary": "Get doctor by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Doctor details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Doctor" }
              }
            }
          }
        }
      }
    },
    "/staff": {
      "post": {
        "summary": "Create a new staff (staff only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Staff" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Staff created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/appointments": {
      "get": {
        "summary": "Get appointments (filtered by role)",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of appointments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Appointment" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new appointment",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Appointment" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/appointments/{id}": {
      "get": {
        "summary": "Get appointment by ID (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Appointment" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update appointment (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Appointment" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Appointment updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Appointment" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete appointment (BOLA vulnerable)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/prescriptions": {
      "post": {
        "summary": "Create a new prescription (doctor only)",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Prescription" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prescription created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Patient": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "date_of_birth": { "type": "string" },
          "phone": { "type": "string" },
          "address": { "type": "string" }
        },
        "required": ["name", "date_of_birth", "phone", "address"]
      },
      "Doctor": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "specialty": { "type": "string" }
        },
        "required": ["name", "specialty"]
      },
      "Staff": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }
        },
        "required": ["name"]
      },
      "Appointment": {
        "type": "object",
        "properties": {
          "patient_id": { "type": "integer" },
          "doctor_id": { "type": "integer" },
          "appointment_time": { "type": "string" },
          "status": { "type": "string" }
        },
        "required": ["patient_id", "doctor_id", "appointment_time", "status"]
      },
      "Prescription": {
        "type": "object",
        "properties": {
          "patient_id": { "type": "integer" },
          "doctor_id": { "type": "integer" },
          "medications": { "type": "string" },
          "date_issued": { "type": "string" }
        },
        "required": ["patient_id", "doctor_id", "medications", "date_issued"]
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}